+++ http/orders.http
@@ -80,6 +80,10 @@ Content-Type: application/json
 ### 주문 취소
 PUT localhost:10366/api/orders/cancelled/1
 
+### 주문 목록 조회
+GET localhost:10366/api/admin/orders?status=pending
+
+
 
 
 
@@ -97,7 +101,7 @@ GET localhost:10366/api/orders/product/list/1
 GET localhost:10366/api/orders/package
 
 ### package 추가
-POST localhost:10366/api/orders/package
+POST localhost:10366/api/admin/orders/package
 Content-Type: application/json
 
 {
@@ -106,7 +110,7 @@ Content-Type: application/json
 }
 
 ### package 수정
-PUT localhost:10366/api/orders/package/1
+PUT localhost:10366/api/admin/orders/package/1
 Content-Type: application/json
 
 {
