+++ src/main/java/shop/sajotuna/order/orders/controller/PackageController.java
@@ -1,10 +1,8 @@
 package shop.sajotuna.order.orders.controller;
 
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import shop.sajotuna.order.orders.dto.PackageRequest;
 import shop.sajotuna.order.orders.dto.PackageResponse;
 import shop.sajotuna.order.orders.service.PackageService;
 
@@ -21,29 +19,5 @@ public class PackageController {
     public ResponseEntity<List<PackageResponse>> getPackage(){
         List<PackageResponse> packages = packageService.getPackages();
         return ResponseEntity.ok(packages);
-    }
-
-    // 포장 생성 (관리자 전용)
-    @PostMapping
-    public ResponseEntity<PackageResponse> createPackage(@RequestBody @Valid PackageRequest request) {
-        PackageResponse packageResponse = packageService.createPackage(request);
-
-        return ResponseEntity.ok(packageResponse);
-    }
-
-    // 포장 수정 (관리자 전용)
-    @PutMapping("/{package-id}")
-    public ResponseEntity<Void> updatePackage(@PathVariable("package-id") Long packageId, @RequestBody @Valid PackageRequest request) {
-        packageService.updatePackage(packageId, request);
-
-        return ResponseEntity.noContent().build();
-    }
-
-    // 포장 삭제 (관리자 전용)
-    @DeleteMapping("/{package-id}")
-    public ResponseEntity<Void> deletePackage(@PathVariable("package-id") Long packageId){
-        packageService.deletePackage(packageId);
-
-        return ResponseEntity.noContent().build();
     }
 }
